完善平台分析

+++ geek_analysis/src/main/java/cn/geekview/entity/model/TdreamWebsite.java
@@ -30,7 +30,7 @@ public class TdreamWebsite {
 
     private Integer totalProducts;
 
-    private Double averageFinish;
+    private BigDecimal averageFinish;
 
     private BigDecimal amountIncreaseDay;
